pyproject.toml
[project]
name = "minor-pj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "tensorflow>=2.19.0",
]
